products

+++ app/Http/Controllers/ProductController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\LangController;
 use App\Models\Brand;
-use App\Models\Product;
 use App\Models\Category;
+use App\Models\Product;
 use App\Models\ProductImage;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use App\Http\Controllers\LangController;
+use Illuminate\Support\Facades\DB;
 
 class ProductController extends Controller
 {
@@ -19,22 +20,44 @@ class ProductController extends Controller
      */
     public function index()
     {
+        $user= Auth::user();
+        $tab = request('tab');
+        $title = "المنتجات";
         if (request()->ajax()) {
             $products = Product::
-            leftJoin('categories as m','m.id','main_category')
-            ->leftJoin('categories as s','s.id','sub_category')
-            ->leftJoin('brands','brands.id','brand_id')
-            ->select('price','qty','m.name_' . LangController::lang() . ' as main_category_name','s.name_' . LangController::lang() . ' as sub_category_name','products.id', 'status', 'products.name_' . LangController::lang() . ' as name', 'products.description_' . LangController::lang() . ' as description','brands.name_' . LangController::lang() .' as brand')
-            ->get()
-            ;
+            where('user_id',$user->id)
+                ->select('duration', DB::raw('date(products.updated_at) as update_date'), 'price', 'qty', 'products.id', 'status', 'products.name_' . LangController::lang() . ' as name', 'products.description_' . LangController::lang() . ' as description')
+                ->orderBy('products.id', 'desc');
+
+            switch ($tab) {
+                case 1:
+                    $products->where('status', 1);
+                    break;
+                case 2:
+                    $products->where('status', 0);
+                    break;
+                case 3:
+                    $products->where('status', 1)->where('qty',0);
+                    break;
+                case 4:
+                    $products->where('status', 2);
+                    break;
+                case 5:
+                    $products->where('status', 3);
+                    break;
+                case 6:
+                    $products->where('status', 4);
+                    break;
+            }
+            $products = $products->get();
 
             $index = 1;
             foreach ($products as $product) {
                 $product->index = $index++;
 
-                $image = ProductImage::where('product_id', $product->id)->first();
+                $image = ProductImage::where('product_id', $product->id)->where('main', 1)->first();
                 if ($image) {
-                    $product->image = asset('uploaded/' .$image->name);
+                    $product->image = asset('uploaded/' . $image->name);
                 } else {
                     $product->image = null;
                 }
@@ -42,7 +65,7 @@ class ProductController extends Controller
             return datatables()->of($products)->addIndexColumn()->make(true);
         }
 
-        return view('vendor.show.products');
+        return view('vendor.show.management', compact('title'));
     }
 
     /**
@@ -52,11 +75,12 @@ class ProductController extends Controller
      */
     public function create()
     {
-        $action='add';
-        $main_categories=Category::whereNull('parent_id')->select('id','name_' . LangController::lang() . ' as name')->get();
-        $sub_categories=Category::whereNotNull('parent_id')->select('id','name_' . LangController::lang() . ' as name','parent_id')->get();
-        $brands=Brand::select('id','name_' . LangController::lang() . ' as name')->get();
-        return view('vendor.add.product',compact('action','main_categories','sub_categories','brands'));
+        $title = "إضافة منتج جديد";
+        $action = 'add';
+        $main_categories = Category::whereNull('parent_id')->select('id', 'name_' . LangController::lang() . ' as name')->get();
+        $sub_categories = Category::whereNotNull('parent_id')->select('id', 'name_' . LangController::lang() . ' as name', 'parent_id')->get();
+        $brands = Brand::select('id', 'name_' . LangController::lang() . ' as name')->get();
+        return view('vendor.add.product', compact('action', 'main_categories', 'sub_categories', 'brands', 'title'));
     }
 
     /**
@@ -67,19 +91,22 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $user= Auth::user();
-        $product=new Product();
-        $product->user_id=$user->id;
-        $product->name_ar=request('name_ar');
-        $product->name_en=request('name_en');
-        $product->description_ar=request('description_ar');
-        $product->description_en=request('description_en');
-        $product->price=request('price');
-        $product->qty=request('qty');
-        $product->main_category=request('main_category');
-        $product->sub_category=request('sub_category');
-        $product->brand_id=request('brand_id');
+        $user = Auth::user();
+        $product = new Product();
+        $product->user_id = $user->id;
+        $product->name_ar = request('name_ar');
+        $product->name_en = request('name_en');
+        $product->duration = request('duration');
+        $product->description_ar = request('description_ar');
+        $product->description_en = request('description_en');
+        $product->price = request('price');
+        $product->qty = request('qty');
+        $product->main_category = request('main_category');
+        $product->sub_category = request('sub_category');
+        $product->brand_id = request('brand_id');
         $product->save();
+
+        $index = 0;
         if ($request->hasfile('images')) {
             foreach ($request->file('images') as $item) {
                 $path = 'uploaded/';
@@ -88,10 +115,14 @@ class ProductController extends Controller
                 $item->move($path, $name);
                 $image->name = $name;
                 $image->product_id = $product->id;
+                if ($index == 0) {
+                    $image->main = 1;
+                }
                 $image->save();
+                $index++;
             }
         }
-        return redirect()->route('product.index')->with('success',__('items.success_add'));
+        return redirect()->route('product.stored')->with('success', __('items.success_add'));
     }
 
     /**
@@ -102,13 +133,14 @@ class ProductController extends Controller
      */
     public function show($id)
     {
-        $action='update';
-        $saved=Product::find($id);
-        $images=ProductImage::where('product_id',$id)->get();
-        $main_categories=Category::whereNull('parent_id')->select('id','name_' . LangController::lang() . ' as name')->get();
-        $sub_categories=Category::whereNotNull('parent_id')->select('id','name_' . LangController::lang() . ' as name','parent_id')->get();
-        $brands=Brand::select('id','name_' . LangController::lang() . ' as name')->get();
-        return view('vendor.add.product',compact('action','main_categories','sub_categories','saved','images','brands'));
+        $title = "تعديل منتج";
+        $action = 'update';
+        $saved = Product::find($id);
+        $images = ProductImage::where('product_id', $id)->get();
+        $main_categories = Category::whereNull('parent_id')->select('id', 'name_' . LangController::lang() . ' as name')->get();
+        $sub_categories = Category::whereNotNull('parent_id')->select('id', 'name_' . LangController::lang() . ' as name', 'parent_id')->get();
+        $brands = Brand::select('id', 'name_' . LangController::lang() . ' as name')->get();
+        return view('vendor.add.product', compact('action', 'main_categories', 'sub_categories', 'saved', 'images', 'brands', 'title'));
     }
 
     /**
@@ -131,16 +163,17 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product=Product::find($id);
-        $product->name_ar=request('name_ar');
-        $product->name_en=request('name_en');
-        $product->description_ar=request('description_ar');
-        $product->description_en=request('description_en');
-        $product->price=request('price');
-        $product->qty=request('qty');
-        $product->main_category=request('main_category');
-        $product->sub_category=request('sub_category');
-        $product->brand_id=request('brand_id');
+        $product = Product::find($id);
+        $product->name_ar = request('name_ar');
+        $product->name_en = request('name_en');
+        $product->duration = request('duration');
+        $product->description_ar = request('description_ar');
+        $product->description_en = request('description_en');
+        $product->price = request('price');
+        $product->qty = request('qty');
+        $product->main_category = request('main_category');
+        $product->sub_category = request('sub_category');
+        $product->brand_id = request('brand_id');
         $product->save();
 
         $images = ProductImage::where('product_id', $product->id)->get();
@@ -150,6 +183,13 @@ class ProductController extends Controller
                 ProductImage::where('id', $image->id)->delete();
             }
         }
+        $first = ProductImage::where('main', 1)->where('product_id', $product->id)->first();
+        if ($first) {
+            $index = 1;
+        } else {
+            $index = 0;
+        }
+
         if ($request->hasfile('images')) {
             foreach ($request->file('images') as $item) {
                 $path = 'uploaded/';
@@ -158,10 +198,22 @@ class ProductController extends Controller
                 $item->move($path, $name);
                 $image->name = $name;
                 $image->product_id = $product->id;
+                if ($index == 0) {
+                    $image->main = 1;
+                }
                 $image->save();
+                $index++;
             }
         }
-        return redirect()->route('product.index')->with('success',__('items.success_updated'));
+        $first_1 = ProductImage::where('main', 1)->where('product_id', $product->id)->first();
+        if (!$first_1) {
+            $first_2 = ProductImage::where('product_id', $product->id)->first();
+            if ($first_2) {
+                $first_2->main = 1;
+                $first_2->save();
+            }
+        }
+        return redirect()->route('product.stored')->with('success', __('items.success_updated'));
     }
 
     /**
@@ -172,16 +224,33 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        Product::find($id)->delete();
-        return redirect()->back()->with('success', __('items.success_delete'));
+        $user = Auth::user();
+        $action = false;
+        if ($user->role_id == 1) {
+            $action = true;
+        } else {
+            $product = Product::where('user_id', $user->id)->where('id', $id)->first();
+            if ($product) {
+                $action = true;
+            }
+        }
+
+        if ($action) {
+            ProductImage::where('product_id', $product->id)->delete();
+            Product::find($id)->delete();
+            return redirect()->back()->with('success', __('items.success_delete'));
+        } else {
+            return redirect()->back();
+        }
     }
 
-    public function change_status($id){
-        $product=Product::find($id);
-        if($product->status==0){
-            $product->status=1;
-        }else{
-            $product->status=0;
+    public function change_status($id)
+    {
+        $product = Product::find($id);
+        if ($product->status == 0) {
+            $product->status = 1;
+        } else {
+            $product->status = 0;
         }
         return redirect()->back()->with('success', __('items.success_update'));
     }
